Refactoring. Removing ProposalModel
