+++ src/models/ProposalSubmissionState.ts
@@ -1,6 +1,21 @@
-import { ProposalSubsetSumbission } from 'models/ProposalModel';
+import { Proposal } from 'generated/sdk';
 
 import { QuestionarySubmissionState } from './QuestionarySubmissionState';
+
+export type ProposalSubsetSumbission = Pick<
+  Proposal,
+  | 'id'
+  | 'abstract'
+  | 'proposer'
+  | 'questionary'
+  | 'status'
+  | 'users'
+  | 'title'
+  | 'shortCode'
+  | 'callId'
+  | 'questionaryId'
+  | 'submitted'
+>;
 
 export interface ProposalSubmissionState extends QuestionarySubmissionState {
   proposal: ProposalSubsetSumbission;
